ai_service/requirements.txt
flask
flask-cors
groq
google-generativeai
pymupdf
pytesseract
Pillow
python-dotenv
jsonschema
requests